Moved CVE reporting into its own runbit.
